+++ src/metadata/metadata_collision.h
@@ -85,10 +85,10 @@ void ocf_metadata_add_to_collision(struct ocf_cache *cache,
 void ocf_metadata_remove_from_collision(struct ocf_cache *cache,
 		ocf_cache_line_t line, ocf_part_id_t part_id);
 
- void ocf_metadata_start_collision_shared_access(
+void ocf_metadata_start_collision_shared_access(
 		struct ocf_cache *cache, ocf_cache_line_t line);
 
- void ocf_metadata_end_collision_shared_access(
+void ocf_metadata_end_collision_shared_access(
 		struct ocf_cache *cache, ocf_cache_line_t line);
 
 #endif /* METADATA_COLLISION_H_ */
